Regressions and regression optimization for all files
In this notebook, we attempt to find the best possible model for virial coefficients of the gases in each experimental dataset. This is ultimately completed by performing 1..n linear regressions and using the Akaike Information Criterion to determine which model provides the best fit. In this notebook, we show that for every gas in each dataset, the best model is provided by performing a single linear regression.