Graphs and constants for ONE linear regression
Included is a PDF with a plot of error metrics (MRE and MAE) for when one linear regression is performed on each gas from every dataset. Furthermore, for each compound in each dataset, we have included CSVs containing fit constants a and b (and error metrics) and PDFs containing the plots of the regressions.